+++ changelog.md
@@ -15,6 +15,7 @@
   none; property.
 * Fail hard when the web-frontend can't reach the backend because of a network error.
 * Use UTC time in the date picker.
+* Refactored handler get_* methods so that they never check for permissions.
 
 ## Released (2021-02-04)